# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Locket Clone API Contract
  description: "Tài liệu đặc tả API chuẩn OpenAPI 3.0 / Swagger UI dành cho Frontend\
    \ (Web & Mobile). Cung cấp đầy đủ các endpoint xác thực (JWT), chia sẻ ảnh (Locket),\
    \ quản lý giao dịch tài chính, kết bạn và lưu trữ ảnh Garage S3."
  contact:
    name: Development Team
    email: dev@locketclone.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Development Server
security:
- Bearer Authentication: []
...
paths:
...
  /api/v1/users/me/settings/personal-info:
    get:
      tags:
      - Users
      summary: Lấy thông tin cá nhân trong Settings
      description: "Lấy thông tin họ tên, username, avatar của người dùng hiện tạ\
        i."
      operationId: getCurrentUserPersonalInfo
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    patch:
      tags:
      - Users
      summary: Cập nhật thông tin cá nhân
      description: "Cập nhật họ, tên, username hoặc link avatar trong phần Cài đặ\
        t."
      operationId: updateCurrentUserPersonalInfo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonalInfoRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/users/me/settings/avatar:
    patch:
      tags:
      - Users
      summary: Tải lên và cập nhật ảnh đại diện mới
      description: "Upload file ảnh đại diện (JPEG/PNG/WEBP). Ảnh tự động được né\
        n, tạo thumbnail và lưu trữ an toàn trên Garage S3."
      operationId: updateCurrentUserAvatar
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/users/me/profile:
    patch:
      tags:
      - Users
      summary: Cập nhật hồ sơ Onboarding
      description: "Cập nhật họ tên, username trong bước onboarding sau khi đăng ký\
        \ tài khoản."
      operationId: updateCurrentUserProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: Lấy thông tin người dùng theo ID
      description: Xem thông tin cơ bản của một người dùng theo UUID.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        isVerified:
          type: boolean
        isGoldMember:
          type: boolean
        profileCompleted:
          type: boolean
...
    UpdatePersonalInfoRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 0
          pattern: ^(?!\s*$).+
        firstName:
          type: string
          maxLength: 50
          minLength: 0
          pattern: ^(?!\s*$).+
        lastName:
          type: string
          maxLength: 50
          minLength: 0
          pattern: ^(?!\s*$).+
        avatarUrl:
          type: string
          maxLength: 255
          minLength: 0
          pattern: ^(?!\s*$).+
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 0
          pattern: ^(?!\s*$).+
        firstName:
          type: string
          maxLength: 50
          minLength: 0
          pattern: ^(?!\s*$).+
        lastName:
          type: string
          maxLength: 50
          minLength: 0
          pattern: ^(?!\s*$).+
        avatarUrl:
          type: string
          maxLength: 255
          minLength: 0
          pattern: ^(?!\s*$).+
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Nhập Access Token (JWT) được trả về từ API /api/v1/auth/login hoặc
        /api/v1/auth/verify
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT